setup layout with warning and move activeTab to overview component
