+++ libs/asset-editor/src/lib/components/asset-editor-page/asset-editor-page.component.html
@@ -14,9 +14,15 @@
   </button>
 </asset-sg-page-header>
 <div style="display: flex">
-  <asset-sg-editor-navigation [assetId]="assetId"></asset-sg-editor-navigation>
+  <asset-sg-editor-navigation [assetId]="assetId" [activeTab]="activeTab"></asset-sg-editor-navigation>
   <div>
-    <asset-sg-editor-tab-page></asset-sg-editor-tab-page>
+    @switch (activeTab) { @case (Tab.General) {
+    <span>{{ activeTab }}</span>
+    } @case (Tab.Administration) {
+    <span>{{ activeTab }}</span>
+    } @default {
+    <span>other</span>
+    } }
     <asset-sg-editor-save></asset-sg-editor-save>
   </div>
 </div>
