+++ libs/asset-editor/src/lib/components/asset-editor-page/asset-editor-page.component.ts
@@ -1,16 +1,17 @@
 import { Component, inject, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { ActivatedRoute, ParamMap } from '@angular/router';
+import { ActivatedRoute, NavigationEnd, ParamMap, Router } from '@angular/router';
 import { RoutingService } from '@asset-sg/client-shared';
 import { ORD } from '@asset-sg/core';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { Store } from '@ngrx/store';
 import * as O from 'fp-ts/Option';
-import { map, Subscription, take } from 'rxjs';
+import { filter, map, startWith, Subscription, take, tap } from 'rxjs';
 import { AssetEditDetailVM } from '../../models';
 import * as actions from '../../state/asset-editor.actions';
 import { AppStateWithAssetEditor } from '../../state/asset-editor.reducer';
 import * as fromAssetEditor from '../../state/asset-editor.selectors';
+import { Tab } from '../asset-editor-navigation';
 
 @UntilDestroy()
 @Component({
@@ -23,8 +24,10 @@ export class AssetEditorPageComponent implements OnInit, OnDestroy {
   public assetId = 'new';
   public asset: AssetEditDetailVM | null = null;
   public form?: FormGroup;
+  public activeTab: Tab = Tab.General;
   private readonly store = inject(Store<AppStateWithAssetEditor>);
   private readonly route = inject(ActivatedRoute);
+  private readonly router = inject(Router);
   private readonly routingService = inject(RoutingService);
   private readonly fb = inject(FormBuilder);
 
@@ -32,7 +35,20 @@ export class AssetEditorPageComponent implements OnInit, OnDestroy {
 
   private readonly subscriptions: Subscription = new Subscription();
 
+  readonly activeItem$ = this.router.events.pipe(
+    filter((event) => event instanceof NavigationEnd),
+    startWith(() => undefined),
+    tap(() => {
+      const segments = (this.router.getCurrentNavigation() ?? this.router.lastSuccessfulNavigation)?.finalUrl?.root
+        .children?.['primary']?.segments;
+      if (segments !== undefined && segments.length > 0) {
+        this.activeTab = segments.pop()?.path as Tab;
+      }
+    })
+  );
+
   public ngOnInit() {
+    this.subscriptions.add(this.activeItem$.subscribe());
     this.loadAssetFromRouteParams();
     // TODO: fix with new endpoints/ data schema for asset
     this.subscriptions.add(
@@ -69,4 +85,6 @@ export class AssetEditorPageComponent implements OnInit, OnDestroy {
       })
     );
   }
+
+  protected readonly Tab = Tab;
 }
