+++ libs/asset-editor/src/lib/components/asset-editor-save/asset-editor-save.component.html
@@ -3,7 +3,6 @@
   <svg-icon key="cancel"></svg-icon>
   <span translate>edit.unsavedChanges</span>
 </div>
-
 }
 <div style="flex: 1"></div>
 <button asset-sg-secondary (click)="discardChanges()" translate>
