+++ libs/asset-editor/src/lib/components/asset-editor-save/asset-editor-save.component.scss
@@ -1,6 +1,9 @@
 @use "../../styles/variables";
 
 :host {
+  position: absolute;
+  bottom: 0;
+  width: calc(100vw - 280px);
   height: 56px;
   padding: 0 10px;
   display: flex;
